requirements.txt
langchain
python-dotenv~=1.1.1
ipykernel
langchain_groq
langchain_google_genai
langchain_groq
langchain_community
pypdf
faiss-cpu
structlog~=25.4.0
pyMuPDF
langchain-openai
pandas

pydantic~=2.11.7
pyyaml~=6.0.2
